refactor: update drops2
