+++ workflow_examples/drops2_example.py
@@ -1,6 +1,6 @@
 from door.data_sources import DROPS2Downloader
 
-from door.utils.time import TimeRange
+from door.tools.timestepping import TimeRange
 from door.utils.space import BoundingBox
 from door.utils import log
